Componetiza botoes de explorar
